Move Dart & UniFFI loggers back to their respective places
