+++ libs/Cargo.toml
@@ -20,6 +20,7 @@ version = "0.6.2"
 aes = "0.8"
 anyhow = { version = "1.0.79", features = ["backtrace"] }
 base64 = "0.13.0"
+env_logger = "0.10"
 bitcoin = "=0.29.2" # Same version as used in gl-client
 # Pin the reqwest dependency until macOS linker issue is fixed: https://github.com/seanmonstar/reqwest/issues/2006
 hex = "0.4"
