+++ libs/sdk-bindings/Cargo.toml
@@ -17,6 +17,7 @@ crate-type = ["staticlib", "cdylib", "lib"]
 anyhow = { workspace = true }
 breez-sdk-core = { path = "../sdk-core" }
 sdk-common = { path = "../sdk-common" }
+env_logger = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 uniffi = { version = "0.23.0", features = ["bindgen-tests", "cli"] }
