+++ libs/sdk-core/src/logger.rs
@@ -1,19 +1,16 @@
 use crate::models::LevelFilter as BindingLevelFilter;
-use crate::{LogEntry, LogStream};
 use anyhow::{anyhow, Result};
 use env_logger::{Builder, Logger, Target};
-use flutter_rust_bridge::StreamSink;
-use lazy_static::lazy_static;
+
 use log::{
-    max_level, set_boxed_logger, set_max_level, warn, LevelFilter, Log, Metadata, Record,
+    max_level, set_boxed_logger, set_max_level, LevelFilter, Log, Metadata, Record,
     STATIC_MAX_LEVEL,
 };
-use parking_lot::RwLock;
+
 use std::fs::OpenOptions;
 use std::io::Write;
 
 use chrono::Utc;
-use std::sync::Once;
 
 /* env_logger */
 
@@ -57,7 +54,7 @@ vls_protocol_signer=warn,
 vls_protocol_signer::handler::HandlerBuilder::do_build=warn
 "#;
 
-fn init_env_logger(target: Option<Target>, filter_level: Option<LevelFilter>) -> Logger {
+pub fn init_env_logger(target: Option<Target>, filter_level: Option<LevelFilter>) -> Logger {
     let mut binding = Builder::new();
     let builder = binding
         .parse_filters(ENV_LOGGER_FILTER)
@@ -79,140 +76,6 @@ fn init_env_logger(target: Option<Target>, filter_level: Option<LevelFilter>) ->
         builder.filter_level(filter_level);
     }
     builder.build()
-}
-
-/* Dart */
-
-static INIT_DART_LOGGER: Once = Once::new();
-
-pub fn init_dart_logger(filter_level: Option<BindingLevelFilter>) {
-    INIT_DART_LOGGER.call_once(|| {
-        let filter_level = get_filter_level(filter_level);
-
-        assert!(
-            filter_level <= STATIC_MAX_LEVEL,
-            "Should respect STATIC_MAX_LEVEL={:?}, which is done in compile time. level{:?}",
-            STATIC_MAX_LEVEL,
-            filter_level
-        );
-
-        let env_logger = init_env_logger(Some(Target::Stdout), Some(filter_level));
-
-        let dart_logger = DartLogger { env_logger };
-        set_boxed_logger(Box::new(dart_logger))
-            .unwrap_or_else(|_| error!("Log stream already created."));
-        set_max_level(filter_level);
-    });
-}
-
-lazy_static! {
-    static ref DART_LOGGER_STREAM_SINK: RwLock<Option<StreamSink<LogEntry>>> = RwLock::new(None);
-}
-
-pub struct DartLogger {
-    env_logger: Logger,
-}
-
-impl DartLogger {
-    pub fn set_stream_sink(stream_sink: StreamSink<LogEntry>) {
-        let mut guard = DART_LOGGER_STREAM_SINK.write();
-        if guard.is_some() {
-            warn!(
-                "BindingLogger::set_stream_sink but already exist a sink, thus overriding. \
-                (This may or may not be a problem. It will happen normally if hot-reload Flutter app.)"
-            );
-        }
-        *guard = Some(stream_sink);
-        drop(guard)
-    }
-
-    fn record_to_entry(record: &Record) -> LogEntry {
-        LogEntry {
-            line: format!("{}", record.args()),
-            level: format!("{}", record.level()),
-        }
-    }
-}
-
-impl Log for DartLogger {
-    fn enabled(&self, metadata: &Metadata) -> bool {
-        metadata.level() <= max_level()
-    }
-
-    fn log(&self, record: &Record) {
-        if self.env_logger.enabled(record.metadata()) {
-            let entry = Self::record_to_entry(record);
-            if let Some(sink) = &*DART_LOGGER_STREAM_SINK.read() {
-                sink.add(entry);
-            }
-        }
-    }
-
-    fn flush(&self) {}
-}
-
-/* UniFFI */
-
-static INIT_UNIFFI_LOGGER: Once = Once::new();
-
-pub fn init_uniffi_logger(
-    log_stream: Box<dyn LogStream>,
-    filter_level: Option<BindingLevelFilter>,
-) {
-    INIT_UNIFFI_LOGGER.call_once(|| {
-        UniFFILogger::set_log_stream(log_stream, filter_level);
-    });
-}
-
-pub struct UniFFILogger {
-    env_logger: env_logger::Logger,
-    log_stream: Box<dyn LogStream>,
-}
-
-impl UniFFILogger {
-    fn set_log_stream(log_stream: Box<dyn LogStream>, filter_level: Option<BindingLevelFilter>) {
-        let filter_level = get_filter_level(filter_level);
-        assert!(
-            filter_level <= STATIC_MAX_LEVEL,
-            "Should respect STATIC_MAX_LEVEL={:?}, which is done in compile time. level{:?}",
-            STATIC_MAX_LEVEL,
-            filter_level
-        );
-
-        let env_logger = init_env_logger(Some(Target::Stdout), Some(filter_level));
-
-        let uniffi_logger = UniFFILogger {
-            env_logger,
-            log_stream,
-        };
-        set_boxed_logger(Box::new(uniffi_logger))
-            .unwrap_or_else(|_| error!("Log stream already created."));
-        set_max_level(filter_level);
-    }
-
-    fn record_to_entry(record: &Record) -> LogEntry {
-        LogEntry {
-            line: format!("{}", record.args()),
-            level: format!("{}", record.level()),
-        }
-    }
-}
-
-impl Log for UniFFILogger {
-    fn enabled(&self, metadata: &Metadata) -> bool {
-        // ignore the internal uniffi log to prevent infinite loop.
-        return metadata.level() <= max_level()
-            && *metadata.target() != *"breez_sdk_bindings::uniffi_binding";
-    }
-
-    fn log(&self, record: &Record) {
-        let metadata = record.metadata();
-        if self.enabled(metadata) && self.env_logger.enabled(metadata) {
-            let entry = Self::record_to_entry(record);
-            self.log_stream.log(entry);
-        }
-    }
-    fn flush(&self) {}
 }
 
 /* Rust */
@@ -310,7 +173,7 @@ impl Log for GlobalSdkLogger {
 
 /* Binding LevelFilter */
 
-fn get_filter_level(filter_level: Option<BindingLevelFilter>) -> LevelFilter {
+pub fn get_filter_level(filter_level: Option<BindingLevelFilter>) -> LevelFilter {
     match filter_level.unwrap_or(BindingLevelFilter::Trace) {
         BindingLevelFilter::Off => LevelFilter::Off,
         BindingLevelFilter::Error => LevelFilter::Error,
